chore(sandcastle): rename plan.ts → issue.ts
The file used to host a Plan schema and parsePlan extractor for the
subset selector's <plan>{...}</plan> output. Once the subset selector
was removed, only the Issue schema and makeBranch helper remained, so
the "plan" name was no longer accurate.

Rename via git mv to preserve history; update three importers
(eligibility.ts, pipeline.ts, the test file itself) plus two stale
comments that still referenced "plan.ts's Issue schema".
